LINK-38 Added kong-utils to extract kong's consumer create and remove
