+++ lib/models/user.js
@@ -1,36 +1,12 @@
 'use strict';
 
 const mongoose  = require('mongoose');
-const request   = require('request-promise');
 const sodium    = require('sodium').api;
 const moment    = require('moment');
 const logger    = require('../logger');
 const crypto    = require('crypto');
 const base64url = require('base64url');
-
-function addConsumer(kongUrl, username, id) {
-    return request({
-        method: 'POST',
-        uri:    `${kongUrl}/consumers`,
-        body:   {
-            username:  username,
-            custom_id: id
-        },
-        json: true
-    });
-}
-
-function addJWTCredentials(kongUrl, username, id, secret) {
-    return request({
-        method: 'POST',
-        uri:    `${kongUrl}/consumers/${username}/jwt`,
-        body:   {
-            key:    id,
-            secret: secret
-        },
-        json: true
-    });
-}
+const kongUtils = require('../kong-utils');
 
 const UserSchema = new mongoose.Schema({
     username: {
@@ -82,10 +58,10 @@ UserSchema.pre('save', function(next) {
 
 UserSchema.post('save', function(doc, next) {
     if (doc.wasNew) {
-        addConsumer(process.env.EVENTOL_AUTH_KONG_URL, doc.username, doc._id)
+        kongUtils.addConsumer(process.env.EVENTOL_AUTH_KONG_URL, doc.username, doc._id)
             .then((res) => {
                 logger.info('Consumer created: ', res);
-                return addJWTCredentials(process.env.EVENTOL_AUTH_KONG_URL, doc.username,
+                return kongUtils.addJWTCredentials(process.env.EVENTOL_AUTH_KONG_URL, doc.username,
                     doc._id, doc.secret);
             })
             .then((res) => {
@@ -94,6 +70,10 @@ UserSchema.post('save', function(doc, next) {
             });
     }
     next();
+});
+
+UserSchema.post('delete', function(doc) {
+    kongUtils.deleteConsumer(process.env.EVENTOL_AUTH_KONG_URL, doc.username);
 });
 
 UserSchema.statics.login = function login(username, password) {
